+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fixed: Correctly traverse `mod` statements within package top source files that are not named `lib.rs` or `main.rs`, by following the `path` setting of each target within the manifest.
+
 - Improved: Don't generate function mutants that have the same AST as the code they're replacing.
 
 ## 23.12.0