+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Improved progress bars and console output, including putting the outcome of each mutant on the left, and the overall progress bar at the bottom.
+- Improved progress bars and console output, including putting the outcome of each mutant on the left, and the overall progress bar at the bottom. Improved display of estimated remaining time, and other times.
 
 - Improved: Don't generate function mutants that have the same AST as the code they're replacing.